added rules for participants and project managers

+++ src/components/Members/index.tsx
@@ -10,21 +10,27 @@ import EmptyList from '../errors/listError/EmptyList';
 import { MemberRole } from '@/store/slices/types/userSliceTypes';
 import { updateMemberRole } from '@/api/requests';
 import { updateRole } from '@/store/slices/projectSlice';
+import { useAuth } from '@/hooks/useAuth';
 
 function Members() {
   const { ownProjects, foreignProjects } = useAppSelector((state) => state.project);
   const [currentProject, setCurrentProject] = useState<Project>();
   const { projectId, list } = useParams();
+  const { id: userId } = useAuth();
   const dispatch = useAppDispatch();
 
   useEffect(() => {
     if (projectId) {
       if (list === 'own') {
         const project = ownProjects.find((project) => project.id === +projectId);
-        setCurrentProject(project);
+        if (project) {
+          setCurrentProject(project);
+        }
       } else {
         const project = foreignProjects.find((project) => project.id === +projectId);
-        setCurrentProject(project);
+        if (project) {
+          setCurrentProject(project);
+        }
       }
     }
   }, [projectId, ownProjects]);
@@ -52,36 +58,40 @@ function Members() {
       <div className={styles.participantsProjectWrapper}>
         {currentProject?.members ? (
           <>
-            {currentProject?.members.map(({ id, email, firstName, lastName, photo, role }) => (
-              <div key={id} className={styles.participantsProjectItem}>
-                <div className={styles.memberInfoWrapper}>
-                  <img className={styles.memberAvatar} src={photo ?? profileIcon} alt='' />
-                  <div>
-                    <h4 className={styles.participantsName}>
-                      {firstName} {lastName}
-                    </h4>
-                    <h3 style={{ textAlign: 'center' }}>
-                      {role === MemberRole.PROJECT_MEMBER ? 'Member' : 'Manager'}
-                    </h3>
+            {currentProject?.members.map(({ id, email, firstName, lastName, photo, role }) => {
+              if (id !== userId) {
+                return (
+                  <div key={id} className={styles.participantsProjectItem}>
+                    <div className={styles.memberInfoWrapper}>
+                      <img className={styles.memberAvatar} src={photo ?? profileIcon} alt='' />
+                      <div>
+                        <h4 className={styles.participantsName}>
+                          {firstName} {lastName}
+                        </h4>
+                        <h3 style={{ textAlign: 'center' }}>
+                          {role === MemberRole.PROJECT_MEMBER ? 'Member' : 'Manager'}
+                        </h3>
+                      </div>
+                    </div>
+                    <p className={styles.participantsEmail}>Email: {email}</p>
+                    {list === 'own' && (
+                      <div className={styles.btnConteiner}>
+                        <button
+                          type='button'
+                          className={styles.messageBtn}
+                          onClick={() => setMemberRole(id, role)}
+                        >
+                          {role === MemberRole.PROJECT_MEMBER ? 'Appoint a Manager' : 'Appoint a Member'}
+                        </button>
+                        <button onClick={() => deleteMember(id)} type='button' className={styles.deleteBtn}>
+                          Delete
+                        </button>
+                      </div>
+                    )}
                   </div>
-                </div>
-                <p className={styles.participantsEmail}>Email: {email}</p>
-                {list === 'own' && (
-                  <div className={styles.btnConteiner}>
-                    <button
-                      type='button'
-                      className={styles.messageBtn}
-                      onClick={() => setMemberRole(id, role)}
-                    >
-                      {role === MemberRole.PROJECT_MEMBER ? 'Appoint a Manager' : 'Appoint a Member'}
-                    </button>
-                    <button onClick={() => deleteMember(id)} type='button' className={styles.deleteBtn}>
-                      Delete
-                    </button>
-                  </div>
-                )}
-              </div>
-            ))}
+                );
+              }
+            })}
           </>
         ) : (
           <EmptyList>Members not found</EmptyList>
